obrazarna-klient/requirements.txt
pygame==2.1.2
picamera==1.13
Pillow==9.1.0

